Move mock data classes to separate utils file
