+++ hi-ml-histopathology/testhisto/testhisto/datamodules/test_slides_datamodule.py
@@ -3,44 +3,13 @@
 #  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 #  ------------------------------------------------------------------------------------------
 import pytest
-import pandas as pd
 import numpy as np
 
-from pathlib import Path
-from typing import Tuple, Union, Optional
 from health_ml.utils.common_utils import is_gpu_available
-
-
-from histopathology.datamodules.base_module import SlidesDataModule
-from histopathology.datasets.base_dataset import SlidesDataset
+from testhisto.utils.utils_slides_datamodule import MockSlidesDataModule
 from testhisto.utils.utils_testhisto import full_ml_test_data_path
 
 no_gpu = not is_gpu_available()
-
-
-class MockSlidesDataset(SlidesDataset):
-    SLIDE_ID_COLUMN = "image_id"
-    IMAGE_COLUMN = "image"
-    LABEL_COLUMN = "isup_grade"
-
-    METADATA_COLUMNS = ("data_provider", "isup_grade", "gleason_score")
-
-    def __init__(
-        self,
-        root: Union[str, Path],
-        dataset_csv: Optional[Union[str, Path]] = None,
-        dataset_df: Optional[pd.DataFrame] = None,
-    ) -> None:
-        super().__init__(root, dataset_csv, dataset_df, validate_columns=False)
-        slide_ids = self.dataset_df.index
-        self.dataset_df[self.IMAGE_COLUMN] = slide_ids + ".tiff"
-        self.validate_columns()
-
-
-class MockSlidesDataModule(SlidesDataModule):
-
-    def get_splits(self) -> Tuple[MockSlidesDataset, MockSlidesDataset, MockSlidesDataset]:
-        return (MockSlidesDataset(self.root_path), MockSlidesDataset(self.root_path), MockSlidesDataset(self.root_path))
 
 
 @pytest.mark.skipif(no_gpu, reason="Test requires GPU")
